refactor: 라이다 adapter 위치 정리
